// src/api/models/ritual-step.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface RitualStep
 */
export interface RitualStep {
    /**
     * 
     * @type {string}
     * @memberof RitualStep
     */
    'title'?: string;
    /**
     * 
     * @type {string}
     * @memberof RitualStep
     */
    'description'?: string;
    /**
     * 
     * @type {number}
     * @memberof RitualStep
     */
    'order'?: number;
    /**
     * 
     * @type {number}
     * @memberof RitualStep
     */
    'durationMinutes'?: number;
    /**
     * 
     * @type {boolean}
     * @memberof RitualStep
     */
    'optional'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof RitualStep
     */
    'materials'?: string;
    /**
     * 
     * @type {string}
     * @memberof RitualStep
     */
    'tips'?: string;
}